convert LruCache to an interface
